Initial Working build
Suffers from a fatal case of Works on My Machine...
But otherwise compiles for Windows 11 and appears to run with most
features enabled.

I'm a sucker for static linking by default so the only thing not here is
libpng and libjpg which are needed, but this will run in a Quake install
from stream without issues.